+++ libft/convert/ft_utoa_base.c
@@ -6,7 +6,7 @@
 /*   By: zraunio <zraunio@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/05 12:58:20 by zraunio           #+#    #+#             */
-/*   Updated: 2021/01/11 10:25:09 by zraunio          ###   ########.fr       */
+/*   Updated: 2021/03/04 14:36:01 by zraunio          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,11 +20,11 @@ static int		ft_base(int nb)
 		return (nb + '0');
 }
 
-char			*ft_utoa_base(unsigned long int nb, unsigned long int base)
+char			*ft_utoa_base(unsigned long int nb, long int base)
 {
 	int						i;
 	char					*res;
-	unsigned long int		tmp;
+	long long int			tmp;
 
 	i = 0;
 	tmp = nb;
