+++ libft/math/dig_count_base.c
@@ -6,13 +6,13 @@
 /*   By: zraunio <zraunio@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/07 10:56:14 by zraunio           #+#    #+#             */
-/*   Updated: 2021/03/01 13:54:15 by zraunio          ###   ########.fr       */
+/*   Updated: 2021/03/04 14:34:00 by zraunio          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-int	dig_count_base(unsigned long int nb, unsigned long int base)
+int		dig_count_base(long long int nb, long int base)
 {
 	size_t	i;
